+++ custom_components/govee_ble_hci/__init__.py
@@ -1,2 +1,2 @@
 """Govee Thermometer/Humidity BLE HCI monitor sensor integration."""
-REQUIREMENTS = ["bleson==0.1.6"]
+REQUIREMENTS = ["bleson==0.1.8"]